Rejigger our peano number implementation, document it
`Succ<Never>` is a really poor type to have appear in error messages.
Since we never care about the precise value unless that value is 1, I've
replaced it with two concrete types to represent one and more than one.
The error message still isn't *great*, but it's at least passable.
